pyproject.toml
[tool.poetry]
name = "currency-data"
version = "0.0.1"
description = ""
authors = ["hk"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
pandas = "^2.1.1"
requests = "^2.31.0"
sqlalchemy = "^2.0.21"
click = "^8.2.0"
boto3 = "^1.28.0"
psycopg2-binary = "^2.9.6"
tabulate = "^0.9.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
coverage = "^7.6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.0"
jupyter = "^1.0.0"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
